carshare_v2/rest_api/serializers.py
from rest_framework import serializers

from carshare_v2.transports.models import Transport


class TransportSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transport
        fields = '__all__'